Moved service classes to better location inside the project
